requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
bcrypt>=4.1.0
PyJWT>=2.8.0
